+++ spinta/datasets/backends/dataframe/backends/xml/components.py
@@ -5,6 +5,7 @@ from spinta.datasets.backends.dataframe.components import DaskBackend
 
 class Xml(DaskBackend):
     type: str = "dask/xml"
+    query_builder_type = "dask/xml"
 
     @contextlib.contextmanager
     def begin(self):