+++ spinta/datasets/backends/dataframe/backends/xml/ufuncs/query/components.py
@@ -1,0 +1,5 @@
+from spinta.datasets.backends.dataframe.ufuncs.query.components import DaskDataFrameQueryBuilder
+
+
+class XmlQueryBuilder(DaskDataFrameQueryBuilder):
+    pass